refactor(gateways): change gateways folder and imports
