+++ src/domain/forum/application/gateways/cryptography/encrypter.ts
