+++ src/domain/forum/application/gateways/cryptography/hash-generator.ts
