+++ src/infra/cryptography/bcrypt.hasher.ts
@@ -1,7 +1,7 @@
 import { compare, hash } from 'bcryptjs';
 
-import { HashComparer } from '@/domain/forum/application/cryptography/hash-comparer';
-import { HashGenerator } from '@/domain/forum/application/cryptography/hash-generator';
+import { HashComparer } from '@/domain/forum/application/gateways/cryptography/hash-comparer';
+import { HashGenerator } from '@/domain/forum/application/gateways/cryptography/hash-generator';
 
 export class BcryptHasher implements HashGenerator, HashComparer {
   private HASH_SALT_LENGTH = 8;
