+++ src/infra/cryptography/cryptography.module.ts
@@ -1,8 +1,8 @@
 import { Module } from '@nestjs/common';
 
-import { Encrypter } from '@/domain/forum/application/cryptography/encrypter';
-import { HashComparer } from '@/domain/forum/application/cryptography/hash-comparer';
-import { HashGenerator } from '@/domain/forum/application/cryptography/hash-generator';
+import { Encrypter } from '@/domain/forum/application/gateways/cryptography/encrypter';
+import { HashComparer } from '@/domain/forum/application/gateways/cryptography/hash-comparer';
+import { HashGenerator } from '@/domain/forum/application/gateways/cryptography/hash-generator';
 
 import { BcryptHasher } from './bcrypt.hasher';
 import { JwtEncrypter } from './jwt.encrypter';
