+++ src/infra/cryptography/jwt.encrypter.ts
@@ -1,7 +1,7 @@
 import { Injectable } from '@nestjs/common';
 import { JwtService } from '@nestjs/jwt';
 
-import { Encrypter } from '@/domain/forum/application/cryptography/encrypter';
+import { Encrypter } from '@/domain/forum/application/gateways/cryptography/encrypter';
 
 @Injectable()
 export class JwtEncrypter implements Encrypter {
