+++ src/infra/storage/storage.module.ts
@@ -1,6 +1,6 @@
 import { Module } from '@nestjs/common';
 
-import { Uploader } from '@/domain/forum/application/storage/uploader';
+import { Uploader } from '@/domain/forum/application/gateways/storage/uploader';
 
 import { EnvModule } from '../env/env.module';
 import { R2Storage } from './r2-storage';
