+++ test/cryptography/fake-encrypter.ts
@@ -1,4 +1,4 @@
-import { Encrypter } from '@/domain/forum/application/cryptography/encrypter';
+import { Encrypter } from '@/domain/forum/application/gateways/cryptography/encrypter';
 
 export class FakeEncrypter implements Encrypter {
   async encrypt(payload: Record<string, unknown>): Promise<string> {
